+++ benches/benches/vm_set/crypto.rs
@@ -114,37 +114,44 @@ pub fn run(c: &mut Criterion) {
     let ed19_secret = ed25519_dalek::SigningKey::generate(&mut rand::rngs::OsRng {});
     let ed19_signature = ed19_secret.sign(&*message);
 
-    run_group_ref(
-        &mut c.benchmark_group("ed19"),
-        "ed19",
-        VmBench::new(op::ed19(0x20, 0x21, 0x22))
-            .with_prepare_script(vec![
-                op::gtf_args(0x20, 0x00, GTFArgs::ScriptData),
-                op::addi(
-                    0x21,
-                    0x20,
+    let mut bench_ed19 = c.benchmark_group("ed19");
+    for i in &linear {
+        bench_ed19.throughput(Throughput::Bytes(*i as u64));
+        run_group_ref(
+            &mut bench_ed19,
+            format!("{i}"),
+            VmBench::new(op::ed19(0x20, 0x21, RegId::ZERO, 0x10))
+                .with_prepare_script(vec![
+                    op::gtf_args(0x20, 0x00, GTFArgs::ScriptData),
+                    op::addi(
+                        0x21,
+                        0x20,
+                        ed19_secret
+                            .verifying_key()
+                            .as_bytes()
+                            .len()
+                            .try_into()
+                            .unwrap(),
+                    ),
+                    op::addi(
+                        0x22,
+                        0x21,
+                        ed19_signature.to_bytes().len().try_into().unwrap(),
+                    ),
+                    op::movi(0x10, *i),
+                    op::cfe(0x10),
+                ])
+                .with_data(
                     ed19_secret
                         .verifying_key()
-                        .as_bytes()
-                        .len()
-                        .try_into()
-                        .unwrap(),
+                        .to_bytes()
+                        .iter()
+                        .chain(ed19_signature.to_bytes().iter())
+                        .chain(message.iter())
+                        .copied()
+                        .collect(),
                 ),
-                op::addi(
-                    0x22,
-                    0x21,
-                    ed19_signature.to_bytes().len().try_into().unwrap(),
-                ),
-            ])
-            .with_data(
-                ed19_secret
-                    .verifying_key()
-                    .to_bytes()
-                    .iter()
-                    .chain(ed19_signature.to_bytes().iter())
-                    .chain(message.iter())
-                    .copied()
-                    .collect(),
-            ),
-    );
+        );
+    }
+    bench_ed19.finish();
 }