+++ crates/fuel-core/src/schema/tx/types.rs
@@ -15,12 +15,14 @@ use crate::{
     },
     query::{
         SimpleBlockData,
+        SimpleTransactionData,
         TransactionQueryData,
     },
     schema::{
         block::Block,
         scalars::{
             AssetId,
+            BlobId,
             Bytes32,
             ContractId,
             HexString,
@@ -53,6 +55,7 @@ use fuel_core_types::{
         field::{
             BytecodeRoot,
             BytecodeWitnessIndex,
+            ChargeableBody,
             InputContract,
             Inputs,
             Maturity,
@@ -86,8 +89,10 @@ use fuel_core_types::{
             TransactionExecutionResult,
             TransactionExecutionStatus,
         },
-        txpool,
-        txpool::TransactionStatus as TxStatus,
+        txpool::{
+            self,
+            TransactionStatus as TxStatus,
+        },
     },
     tai64::Tai64,
 };
@@ -184,6 +189,13 @@ impl SuccessStatus {
         Ok(block.into())
     }
 
+    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    async fn transaction(&self, ctx: &Context<'_>) -> async_graphql::Result<Transaction> {
+        let query = ctx.read_view()?;
+        let transaction = query.transaction(&self.tx_id)?;
+        Ok(Transaction::from_tx(self.tx_id, transaction))
+    }
+
     async fn time(&self) -> Tai64Timestamp {
         Tai64Timestamp(self.time)
     }
@@ -231,6 +243,13 @@ impl FailureStatus {
         let query = ctx.read_view()?;
         let block = query.block(&self.block_height)?;
         Ok(block.into())
+    }
+
+    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    async fn transaction(&self, ctx: &Context<'_>) -> async_graphql::Result<Transaction> {
+        let query = ctx.read_view()?;
+        let transaction = query.transaction(&self.tx_id)?;
+        Ok(Transaction::from_tx(self.tx_id, transaction))
     }
 
     async fn time(&self) -> Tai64Timestamp {
@@ -426,6 +445,11 @@ impl Transaction {
                     .map(|c| AssetId(*c))
                     .collect(),
             ),
+            fuel_tx::Transaction::Blob(tx) => Some(
+                tx.input_asset_ids(base_asset_id)
+                    .map(|c| AssetId(*c))
+                    .collect(),
+            ),
         }
     }
 
@@ -446,6 +470,9 @@ impl Transaction {
             fuel_tx::Transaction::Upload(tx) => {
                 Some(input_contracts(tx).map(|v| (*v).into()).collect())
             }
+            fuel_tx::Transaction::Blob(tx) => {
+                Some(input_contracts(tx).map(|v| (*v).into()).collect())
+            }
         }
     }
 
@@ -454,7 +481,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some(mint.input_contract().into()),
         }
     }
@@ -466,6 +494,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(tx) => Some((*tx.policies()).into()),
             fuel_tx::Transaction::Upload(tx) => Some((*tx.policies()).into()),
+            fuel_tx::Transaction::Blob(tx) => Some((*tx.policies()).into()),
         }
     }
 
@@ -477,7 +506,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -488,6 +518,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(tx) => Some(tx.maturity().into()),
             fuel_tx::Transaction::Upload(tx) => Some(tx.maturity().into()),
+            fuel_tx::Transaction::Blob(tx) => Some(tx.maturity().into()),
         }
     }
 
@@ -496,7 +527,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.mint_amount()).into()),
         }
     }
@@ -506,7 +538,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.mint_asset_id()).into()),
         }
     }
@@ -516,7 +549,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.gas_price()).into()),
         }
     }
@@ -527,7 +561,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.tx_pointer()).into()),
         }
     }
@@ -552,6 +587,10 @@ impl Transaction {
         self.0.is_upload()
     }
 
+    async fn is_blob(&self) -> bool {
+        self.0.is_blob()
+    }
+
     async fn inputs(&self) -> Option<Vec<Input>> {
         match &self.0 {
             fuel_tx::Transaction::Script(tx) => {
@@ -565,6 +604,9 @@ impl Transaction {
                 Some(tx.inputs().iter().map(Into::into).collect())
             }
             fuel_tx::Transaction::Upload(tx) => {
+                Some(tx.inputs().iter().map(Into::into).collect())
+            }
+            fuel_tx::Transaction::Blob(tx) => {
                 Some(tx.inputs().iter().map(Into::into).collect())
             }
         }
@@ -585,6 +627,9 @@ impl Transaction {
             fuel_tx::Transaction::Upload(tx) => {
                 tx.outputs().iter().map(Into::into).collect()
             }
+            fuel_tx::Transaction::Blob(tx) => {
+                tx.outputs().iter().map(Into::into).collect()
+            }
         }
     }
 
@@ -593,7 +638,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some(mint.output_contract().into()),
         }
     }
@@ -625,6 +671,12 @@ impl Transaction {
                     .map(|w| HexString(w.clone().into_inner()))
                     .collect(),
             ),
+            fuel_tx::Transaction::Blob(tx) => Some(
+                tx.witnesses()
+                    .iter()
+                    .map(|w| HexString(w.clone().into_inner()))
+                    .collect(),
+            ),
         }
     }
 
@@ -636,7 +688,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -659,7 +712,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -671,7 +725,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -686,6 +741,14 @@ impl Transaction {
             fuel_tx::Transaction::Upload(tx) => {
                 Some((*tx.bytecode_witness_index()).into())
             }
+            fuel_tx::Transaction::Blob(tx) => Some((*tx.bytecode_witness_index()).into()),
+        }
+    }
+
+    async fn blob_id(&self) -> Option<BlobId> {
+        match &self.0 {
+            fuel_tx::Transaction::Blob(blob) => Some(blob.body().id.into()),
+            _ => None,
         }
     }
 
@@ -696,6 +759,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(_) => None,
             fuel_tx::Transaction::Upload(_) => None,
+            fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -721,6 +785,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(_) => None,
             fuel_tx::Transaction::Upload(_) => None,
+            fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -729,7 +794,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => Some((*tx.bytecode_root()).into()),
         }
     }
@@ -739,7 +805,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => Some((*tx.subsection_index()).into()),
         }
     }
@@ -749,7 +816,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => Some((*tx.subsections_number()).into()),
         }
     }
@@ -759,7 +827,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => {
                 Some(tx.proof_set().iter().map(|proof| (*proof).into()).collect())
             }
@@ -773,6 +842,7 @@ impl Transaction {
             | fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(tx) => Some((*tx.upgrade_purpose()).into()),
             fuel_tx::Transaction::Upload(_) => None,
+            fuel_tx::Transaction::Blob(_) => None,
         }
     }