+++ crates/fuel-core/src/state/rocks_db.rs
@@ -42,6 +42,7 @@ use rocksdb::{
 };
 use std::{
     cmp,
+    collections::BTreeMap,
     env,
     fmt,
     fmt::Formatter,
@@ -313,17 +314,30 @@ where
 
         let existing_column_families = DB::list_cf(&opts, &path).unwrap_or_default();
 
-        let mut cf_descriptors_to_open = vec![];
-        let mut cf_descriptors_to_create = vec![];
+        let mut cf_descriptors_to_open = BTreeMap::new();
+        let mut cf_descriptors_to_create = BTreeMap::new();
         for column in columns.clone() {
             let column_name = Self::col_name(column.id());
             let opts = Self::cf_opts(column, &block_opts);
             if existing_column_families.contains(&column_name) {
-                cf_descriptors_to_open.push((column_name, opts));
+                cf_descriptors_to_open.insert(column_name, opts);
             } else {
-                cf_descriptors_to_create.push((column_name, opts));
+                cf_descriptors_to_create.insert(column_name, opts);
             }
         }
+
+        let unknown_columns_to_open: BTreeMap<_, _> = existing_column_families
+            .iter()
+            .filter(|column_name| {
+                !cf_descriptors_to_open.contains_key(*column_name)
+                    && !cf_descriptors_to_create.contains_key(*column_name)
+            })
+            .map(|unknown_column_name| {
+                let unknown_column_options = Self::default_opts(&block_opts);
+                (unknown_column_name.clone(), unknown_column_options)
+            })
+            .collect();
+        cf_descriptors_to_open.extend(unknown_columns_to_open);
 
         let iterator = cf_descriptors_to_open
             .clone()
@@ -430,11 +444,17 @@ where
         format!("col-{}", column)
     }
 
-    fn cf_opts(column: Description::Column, block_opts: &BlockBasedOptions) -> Options {
+    fn default_opts(block_opts: &BlockBasedOptions) -> Options {
         let mut opts = Options::default();
         opts.create_if_missing(true);
         opts.set_compression_type(DBCompressionType::Lz4);
         opts.set_block_based_table_factory(block_opts);
+
+        opts
+    }
+
+    fn cf_opts(column: Description::Column, block_opts: &BlockBasedOptions) -> Options {
+        let mut opts = Self::default_opts(block_opts);
 
         // All double-keys should be configured here
         if let Some(size) = Description::prefix(&column) {
@@ -760,6 +780,7 @@ fn next_prefix(mut prefix: Vec<u8>) -> Option<Vec<u8>> {
     None
 }
 
+#[allow(non_snake_case)]
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -994,9 +1015,9 @@ mod tests {
         let (_primary_db, tmp_dir) = create_db();
 
         // When
-        let old_columns =
-            vec![Column::Coins, Column::Messages, Column::UploadedBytecodes];
-        let result = RocksDb::<OnChain>::open(tmp_dir.path(), old_columns.clone(), None);
+        let columns = enum_iterator::all::<<OnChain as DatabaseDescription>::Column>()
+            .collect::<Vec<_>>();
+        let result = RocksDb::<OnChain>::open(tmp_dir.path(), columns, None);
 
         // Then
         assert!(result.is_err());
@@ -1114,5 +1135,24 @@ mod tests {
 
         // Then
         drop(snapshot);
+    }
+
+    #[test]
+    fn open__opens_subset_of_columns_after_opening_all_columns() {
+        // Given
+        let (first_open_with_all_columns, tmp_dir) = create_db();
+
+        // When
+        drop(first_open_with_all_columns);
+        let part_of_columns =
+            enum_iterator::all::<<OnChain as DatabaseDescription>::Column>()
+                .skip(1)
+                .collect::<Vec<_>>();
+        let open_with_part_of_columns =
+            RocksDb::<OnChain>::open(tmp_dir.path(), part_of_columns, None);
+
+        // Then
+        let _ = open_with_part_of_columns
+            .expect("Should open the database with shorter number of columns");
     }
 }