+++ crates/services/gas_price_service/src/fuel_gas_price_updater.rs
@@ -190,14 +190,28 @@ where
         target_block_height: BlockHeight,
         l2_block_source: L2,
         metadata_storage: Metadata,
+        min_exec_gas_price: u64,
+        exec_gas_price_change_percent: u64,
+        l2_block_fullness_threshold_percent: u64,
     ) -> Result<Self> {
-        let inner = metadata_storage
-            .get_metadata(&target_block_height)?
-            .ok_or(Error::CouldNotInitUpdater(anyhow::anyhow!(
+        let old_metadata = metadata_storage.get_metadata(&target_block_height)?.ok_or(
+            Error::CouldNotInitUpdater(anyhow::anyhow!(
                 "No metadata found for block height: {:?}",
                 target_block_height
-            )))?
-            .into();
+            )),
+        )?;
+        let inner = match old_metadata {
+            UpdaterMetadata::V0(old) => {
+                let v0 = AlgorithmUpdaterV0::new(
+                    old.new_exec_price,
+                    min_exec_gas_price,
+                    exec_gas_price_change_percent,
+                    old.l2_block_height,
+                    l2_block_fullness_threshold_percent,
+                );
+                AlgorithmUpdater::V0(v0)
+            }
+        };
         let updater = Self {
             inner,
             l2_block_source,